# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CategoryResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.ProductListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProductResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.ProductRequest:
    properties:
      description:
        example: High-performance laptop
        type: string
      name:
        example: Laptop
        type: string
      price:
        example: 999.99
        type: number
      quantity:
        example: 50
        type: integer
    type: object
  dto.ProductResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      updated_at:
        type: string
      variants:
        items:
          $ref: '#/definitions/dto.ProductVariantResponse'
        type: array
    type: object
  dto.ProductVariantListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProductVariantResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.ProductVariantRequest:
    properties:
      price_override:
        description: Optional price override
        example: 99.99
        type: number
      product_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      quantity:
        example: 10
        type: integer
      variant_name:
        example: Color
        type: string
      variant_value:
        example: Red
        type: string
    type: object
  dto.ProductVariantResponse:
    properties:
      created_at:
        type: string
      has_override:
        description: Indicates if price is overridden
        type: boolean
      id:
        type: string
      price:
        description: Effective price (override or base product price)
        type: number
      price_override:
        description: The override value if set
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
      variant_name:
        type: string
      variant_value:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: marco@example.com
    name: API Support
  description: RESTful API for managing products and orders in an e-commerce system
    with payment webhook integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go E-Commerce API
  version: "1.0"
paths:
...
  /product_variants/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific product variant
      parameters:
      - description: Product Variant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a product variant by ID
      tags:
      - product_variants
  /products:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products with optional filtering and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      - default: '"created_at"'
        description: Sort by field (name, price, created_at)
        in: query
        name: sort_by
        type: string
      - default: '"desc"'
        description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      - default: true
        description: Filter products in stock only
        in: query
        name: in_stock_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided information
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product's information
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a product
      tags:
      - products
...
  /products/{id}/variants:
    get:
      consumes:
      - application/json
      description: Get a paginated list of product variants for a specific product
        with sorting
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      - default: '"created_at"'
        description: Sort by field (variant_name, created_at)
        in: query
        name: sort_by
        type: string
      - default: '"asc"'
        description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductVariantListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List all product variants for a product
      tags:
      - product_variants
    post:
      consumes:
      - application/json
      description: Create a new product variant with the provided information. Requires
        admin privileges.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product variant information
        in: body
        name: product_variant
        required: true
        schema:
          $ref: '#/definitions/dto.ProductVariantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - requires product:create permission
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product variant
      tags:
      - product_variants
  /variants/{variant_id}:
    delete:
      consumes:
      - application/json
      description: Delete a product variant by ID. Requires admin privileges.
      parameters:
      - description: Product Variant ID
        in: path
        name: variant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - requires product:delete permission
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product variant
      tags:
      - product_variants
    put:
      consumes:
      - application/json
      description: Update an existing product variant's information. Requires admin
        privileges.
      parameters:
      - description: Product Variant ID
        in: path
        name: variant_id
        required: true
        type: string
      - description: Product Variant information
        in: body
        name: product_variant
        required: true
        schema:
          $ref: '#/definitions/dto.ProductVariantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - requires product:update permission
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product variant
      tags:
      - product_variants
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"